Add files via upload
This is a report that @Riya-Virani and I had created to step by step, walk you through our project and give you a detailed understanding about how we went about building our Recommendation Model and what algorithms we have used to Perform Sentiment Analysis. I strongly suggest you go through this report in order to understand why we did what we did with the data we had with us.